+++ test/test.js
@@ -1,5 +1,5 @@
-import fetch, { FormData } from 'node-fetch'
-import Imageboard, { getCommentText } from 'imageboard'
+import Imageboard from 'imageboard/node'
+import { getCommentText, createHttpRequestFunction } from 'imageboard'
 
 // Usage examples:
 // * `npm run test-chan` — uses `4chan`
@@ -7,87 +7,7 @@ import Imageboard, { getCommentText } from 'imageboard'
 // * etc
 const IMAGEBOARD_ID = process.argv[2] || '4chan'
 
-const imageboard = Imageboard(IMAGEBOARD_ID, {
-  // Sends an HTTP request.
-  // Any HTTP request library can be used here.
-  request: (method, url, { body, headers }) => {
-    // If request "Content-Type" is set to be "multipart/form-data",
-    // convert the `body` object to a `FormData` instance.
-    if (headers['content-type'] === 'multipart/form-data') {
-      body = createFormData(body)
-      // Remove `Content-Type` header so that it autogenerates it from the `FormData`.
-      // Example: "multipart/form-data; boundary=----WebKitFormBoundaryZEglkYA7NndbejbB".
-      delete headers['content-type']
-    } else {
-      body = JSON.stringify(body)
-    }
-    return fetch(url, {
-      method,
-      headers,
-      body,
-      // By default, `fetch()` follows any redirects in the process.
-      // Many imageboards have API endpoints that set cookies and then redirect.
-      // If `fetch()` was to follow those redirects, those `set-cookie` headers
-      // from a `status: 302` response would be ignored, and the `imageboard` library
-      // should be able to inspect those `set-cookie` headers in order to extract
-      // their values. So `fetch()` is specifically configured to not follow redirects.
-      redirect: 'manual'
-    }).then((response) => {
-      if (response.status < 400) {
-        return response.text().then((responseText) => ({
-          // If there were any redirects in the proces,
-          // `response.url` is gonna be the final "redirected to" URL.
-          url: response.url,
-          status: response.status,
-          headers: response.headers,
-          responseText
-        }))
-      }
-      return rejectWithErrorForResponse(response)
-    })
-  }
-})
-
-// Creates an error from a `fetch()` response.
-// Returns a `Promise` and rejects it with the error.
-function rejectWithErrorForResponse(response) {
-  const error = new Error(response.statusText)
-  // If there were any redirects in the proces,
-  // `response.url` is gonna be the final "redirected to" URL.
-  error.url = response.url
-  error.status = response.status
-  error.headers = response.headers
-  return response.text().then(
-    (responseText) => {
-      error.responseText = responseText
-      throw error
-    },
-    (error_) => {
-      throw error
-    }
-  )
-}
-
-// Converts an object to a `FormData` instance.
-function createFormData(body) {
-  // * For 'multipart/form-data', use `FormData` class.
-  // * For 'application/x-www-form-urlencoded', use `URLSearchParams` class.
-  const formData = new FormData()
-  if (body) {
-    for (const key of Object.keys(body)) {
-      if (body[key] !== undefined && body[key] !== null) {
-        if (Array.isArray(body[key])) {
-          for (const element of body[key]) {
-            formData.append(key + '[]', element)
-          }
-        } else {
-          formData.append(key, body[key])
-        }
-      }
-    }
-  }
-  return formData
-}
+const imageboard = Imageboard(IMAGEBOARD_ID)
 
 // // Test `2ch.hk` really old archived thread.
 // return imageboard.getThread({